third_party/WebKit/LayoutTests/imported/wpt/IndexedDB/idbobjectstore_delete2-expected.txt

This is a testharness.js-based test.
FAIL IDBObjectStore.delete() - key doesn't match any records  assert_equals: expected (undefined) undefined but got (number) 0
Harness: the test ran to completion.
